+++ dear-imgui/src/lib.rs
@@ -333,7 +333,9 @@ pub use self::string::*;
 pub use self::style::*;
 pub use self::ui::*;
 // Re-export utility flags/types for convenience
-pub use self::utils::{FocusedFlags, HoveredFlags, LogAutoOpenDepth};
+pub use self::utils::{
+    FocusedFlags, ItemHoveredFlags, LogAutoOpenDepth, TooltipHoveredFlags, WindowHoveredFlags,
+};
 
 // Utility modules
 pub use self::list_clipper::*;